replaced dummy start with actual start methods
